+++ products/speaker/common_components/espressif__esp_lvgl_port/src/lvgl9/esp_lvgl_port.c
@@ -15,6 +15,7 @@
 #include "freertos/task.h"
 #include "freertos/semphr.h"
 #include "freertos/event_groups.h"
+#include "freertos/idf_additions.h"
 #include "esp_lvgl_port.h"
 #include "esp_lvgl_port_priv.h"
 #include "lvgl.h"
@@ -37,8 +38,6 @@ typedef struct lvgl_port_ctx_s {
     bool                running;
     int                 task_max_sleep_ms;
     int                 timer_period_ms;
-    uint8_t             *task_stack_buffer;
-    StaticTask_t        task_stack;
 } lvgl_port_ctx_t;
 
 /*******************************************************************************
@@ -85,19 +84,14 @@ esp_err_t lvgl_port_init(const lvgl_port_cfg_t *cfg)
     lvgl_port_ctx.lvgl_events = xEventGroupCreate();
     ESP_GOTO_ON_FALSE(lvgl_port_ctx.lvgl_events, ESP_ERR_NO_MEM, err, TAG, "Create LVGL Event Group fail!");
 
-    int task_stack_cap = MALLOC_CAP_INTERNAL | MALLOC_CAP_8BIT;
-    if (cfg->task_in_ext) {
-        task_stack_cap = MALLOC_CAP_SPIRAM | MALLOC_CAP_8BIT;
-    }
-    lvgl_port_ctx.task_stack_buffer = (uint8_t *)heap_caps_malloc(cfg->task_stack, task_stack_cap);
-    ESP_GOTO_ON_FALSE(lvgl_port_ctx.task_stack_buffer, ESP_ERR_NO_MEM, err, TAG, "Create LVGL task stack fail!");
-
+    BaseType_t res;
+    const uint32_t caps = cfg->task_stack_caps ? cfg->task_stack_caps : MALLOC_CAP_DEFAULT; // caps cannot be zero
     if (cfg->task_affinity < 0) {
-        lvgl_port_ctx.lvgl_task = xTaskCreateStatic(lvgl_port_task, "taskLVGL", cfg->task_stack, xTaskGetCurrentTaskHandle(), cfg->task_priority, lvgl_port_ctx.task_stack_buffer, &lvgl_port_ctx.task_stack);
+        res = xTaskCreateWithCaps(lvgl_port_task, "taskLVGL", cfg->task_stack, xTaskGetCurrentTaskHandle(), cfg->task_priority, &lvgl_port_ctx.lvgl_task, caps);
     } else {
-        lvgl_port_ctx.lvgl_task = xTaskCreateStaticPinnedToCore(lvgl_port_task, "taskLVGL", cfg->task_stack, xTaskGetCurrentTaskHandle(), cfg->task_priority, lvgl_port_ctx.task_stack_buffer, &lvgl_port_ctx.task_stack, cfg->task_affinity);
+        res = xTaskCreatePinnedToCoreWithCaps(lvgl_port_task, "taskLVGL", cfg->task_stack, xTaskGetCurrentTaskHandle(), cfg->task_priority, &lvgl_port_ctx.lvgl_task, cfg->task_affinity, caps);
     }
-    ESP_GOTO_ON_FALSE(lvgl_port_ctx.lvgl_task, ESP_FAIL, err, TAG, "Create LVGL task fail!");
+    ESP_GOTO_ON_FALSE(res == pdPASS, ESP_FAIL, err, TAG, "Create LVGL task fail!");
 
     // Wait until taskLVGL starts
     if (ulTaskNotifyTake(pdTRUE, pdMS_TO_TICKS(5000)) == 0) {